+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Changes
 - Some libraries were updated.
 - Encryption technical view was improved visually.
+- Compress/Decompress technical view was improved visually.
 
 # 2026.07.15